requirements.txt
Django
jinja2
python-dotenv
json
datetime
tzlocal
pytz
multiprocessing
pymysq
bootstrap4
PyCryptodome
python-dateutil
cryptography